Updated LFeaturesSet in rest of project
